% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgIn.R
\name{pkgInFunc}
\alias{pkgInFunc}
\title{A dummy function that shows linking to native routines in pkgOut works}
\usage{
pkgInFunc()
}
\value{
Side effect is used
}
\description{
A dummy function that shows linking to native routines in pkgOut works
}
